add ros visualizations
