+++ hydra_ros/src/active_window/reconstruction_visualizer.cpp
@@ -50,8 +50,6 @@ using sensor_msgs::msg::Image;
 using std_msgs::msg::ColorRGBA;
 using visualization_msgs::msg::Marker;
 using visualization_msgs::msg::MarkerArray;
-using visualizer::ContinuousPalette;
-using visualizer::DivergentPalette;
 using visualizer::RangeColormap;
 
 namespace {
